clean up/refactor chokidar and mv to public
